requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
Django==3.0.3
django-crispy-forms==1.8.1
gTTS==2.1.1
gTTS-token==1.1.3
idna==2.9
Pillow==7.0.0
pkg-resources==0.0.0
python-decouple==3.3
pytz==2019.3
requests==2.23.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.0
urllib3==1.25.8
